+++ api/EquipmentAllocations/Services/IBookingTransactionalService.cs
@@ -6,6 +6,6 @@ namespace EquipmentAllocations.Services
     public interface IBookingTransactionalService
     {
         Task<BookingDto> IssueBookingAsync(CreateBookingDto dto, string? idempotencyKey);
-        Task<BookingDto> UpdateBookingAsync(int id, UpdateBookingDto dto);
+        Task<BookingDto> UpdateBookingAsync(long id, UpdateBookingDto dto);
     }
 }
